requirements.txt
pydantic>=2.6
numpy>=1.26
tqdm>=4.66
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
streamlit>=1.32
